+++ CHANGELOG.md
@@ -17,6 +17,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 - renamed Honey Fungus to Honey Waxcap #76
 - renamed and reworked Poisonous Mushroom (poisonous, purple, green) to Deadly Fibrecap (poisonous, white, fibre loot) #77
 
+### Fixed
+- banner and bed created from of Slime Fungus aka. Parrot Waxcap were orange instead of lime
+
 ## [1.20.1-4.1.0.0] - 2023-08-09
 ### Changed
 - Changed Forge to NeoForge 1.20.1-47.1.54 (compatible with Forge 47.1.0)